tests/wpt/web-platform-tests/app-history/navigate/state/updateCurrent/state-after-history-replaceState.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
  appHistory.updateCurrent({ state: { data: "value" } });
  assert_equals(appHistory.current.getState().data, "value");
  history.replaceState(1, "", "#replace");
  assert_equals(appHistory.current.getState(), undefined);
}, "appHistoryEntry.getState() after history.replaceState()");
</script>
